Move to perf tests
